requirements.txt
Package                Version
---------------------- ----------
click                  8.1.3
joblib                 1.2.0
nltk                   3.8.1
pip                    22.3.1
py-readability-metrics 1.4.5
regex                  2022.10.31
setuptools             65.5.0
tabulate               0.9.0
tqdm                   4.65.0
